+++ test/interpolation/features/completion/basic.test.ts
@@ -67,7 +67,7 @@ describe('Should autocomplete interpolation for <template>', () => {
     });
 
     it(`completes child component tag`, async () => {
-      await testCompletion(parentTemplateDocUri, position(4, 5), [
+      await testCompletion(parentTemplateDocUri, position(6, 5), [
         {
           label: 'basic',
           kind: CompletionItemKind.Property,
@@ -88,6 +88,26 @@ describe('Should autocomplete interpolation for <template>', () => {
 }`,
             'js'
           )
+        }
+      ]);
+    });
+
+    it(`completes child component's props when kebab case component name`, async () => {
+      await testCompletion(parentTemplateDocUri, position(4, 15), [
+        {
+          label: 'bar',
+          kind: CompletionItemKind.Value,
+          documentation: new MarkdownString('My bar').appendCodeblock(`bar: String`, 'js')
+        }
+      ]);
+    });
+
+    it(`completes child component's props when camel case component name`, async () => {
+      await testCompletion(parentTemplateDocUri, position(5, 14), [
+        {
+          label: 'bar',
+          kind: CompletionItemKind.Value,
+          documentation: new MarkdownString('My bar').appendCodeblock(`bar: String`, 'js')
         }
       ]);
     });